requirements.txt
fastapi[all]
uvicorn
motor
python-dotenv